StringID: Changed constant to avoid warning on clang

+++ lib-core/tests/sfz/strings/StringID_Tests.cpp
@@ -57,7 +57,7 @@ TEST_CASE("Ensuring we always get same hash for same string", "[sfz::StringID]")
 	REQUIRE(collection.numStringsHeld() == 0);
 
 	StringID helloWorldId = collection.getStringID("Hello World!");
-	constexpr uint64_t HELLO_WORLD_HASH = uint64_t(10092224619179044402);
+	constexpr uint64_t HELLO_WORLD_HASH = 10092224619179044402ull;
 	REQUIRE(helloWorldId.id == HELLO_WORLD_HASH);
 
 	// Ensure we get same string id both times
